Add MIT License information for _tag_perceptron.py
