--- pythainlp/tag/_perceptron.py
@@ -1,82 +1,0 @@
-# -*- coding: utf-8 -*-
-"""
-Averaged perceptron classifier.
-Implementation geared for simplicity rather than efficiency.
-
-Code from
-https://github.com/sloria/textblob-aptagger/blob/master/textblob_aptagger/_perceptron.py
-"""
-import pickle
-import random
-from collections import defaultdict
-from typing import Dict
-
-
-class AveragedPerceptron(object):
-
-    """
-    An averaged perceptron, as implemented by Matthew Honnibal.
-
-    See more implementation details here:
-        http://honnibal.wordpress.com/2013/09/11/a-good-part-of-speechpos-tagger-in-about-200-lines-of-python/
-    """
-
-    def __init__(self) -> None:
-        # Each feature gets its own weight vector,
-        # so weights is a dict-of-dicts
-        self.weights = {}
-        self.classes = set()
-        # The accumulated values, for the averaging. These will be keyed by
-        # feature/class tuples
-        self._totals = defaultdict(int)
-        # The last time the feature was changed, for the averaging. Also
-        # keyed by feature/class tuples
-        # (tstamps is short for timestamps)
-        self._tstamps = defaultdict(int)
-        # Number of instances seen
-        self.i = 0
-
-    def predict(self, features: Dict):
-        """
-        Dot-product the features and current weights and return the best
-        label.
-        """
-        scores = defaultdict(float)
-        for feat, value in features.items():
-            if feat not in self.weights or value == 0:
-                continue
-            weights = self.weights[feat]
-            for label, weight in weights.items():
-                scores[label] += value * weight
-        # Do a secondary alphabetic sort, for stability
-        return max(self.classes, key=lambda label: (scores[label], label))
-
-    def update(self, truth, guess, features: Dict) -> None:
-        """Update the feature weights."""
-
-        def upd_feat(c, f, w, v):
-            param = (f, c)
-            self._totals[param] += (self.i - self._tstamps[param]) * w
-            self._tstamps[param] = self.i
-            self.weights[f][c] = w + v
-
-        self.i += 1
-        if truth == guess:
-            return
-        for f in features:
-            weights = self.weights.setdefault(f, {})
-            upd_feat(truth, f, weights.get(truth, 0.0), 1.0)
-            upd_feat(guess, f, weights.get(guess, 0.0), -1.0)
-
-    def average_weights(self) -> None:
-        """Average weights from all iterations."""
-        for feat, weights in self.weights.items():
-            new_feat_weights = {}
-            for clas, weight in weights.items():
-                param = (feat, clas)
-                total = self._totals[param]
-                total += (self.i - self._tstamps[param]) * weight
-                averaged = round(total / float(self.i), 3)
-                if averaged:
-                    new_feat_weights[clas] = averaged
-            self.weights[feat] = new_feat_weights
